Add lazy gurobi as its own solver
